fix: Sets Makefile default target to "all: build"
This patch updates the default target in the Makefile to be "all" which
depends on "build". This is a standard target graph in Makefiles.
Currently the Makefile's default target is "$(GOBIN)" which is a
directory path. For many reasons directory targets do not make the best
Makefile targets *or* dependencies (unless order-only), but this is also
not a user-friendly choice for the default target. Executing "make"
should not yield "nothing to be done" a clean system with an existing
directory "$GOPATH/bin".