Adds CCM user; fixes "watch secret" errors
This patch binds the user "cloud-controller-manager" to the role
"system:cloud-controller-manager" in order to enable the use of a custom
Kubernetes user and thus no longer needing to provide the CCM with the
controller-manager's kubeconfig.

This patch also fixes the annoying "watch secret" failures that populate
the logs. Adding the "watch" permission to the "secrets" resource for
the "system:cloud-controller-manager" role and using the
"cloud-controller-manager" user's kubeconfig will result in these errors
being a thing of the past.