Add Kubernetes e2e conformance testing
This patch adds a conformance testing stage to Travis-CI builds. All
commits to master will cause a new Kubernetes cluster to be turned up on
VMC, run the e2e conformance tests, turn down the cluster, and then the
results will be uploaded to a well-known GCS bucket that is scraped by
test-grid.