Always initialize a Controller Service early
The previous code would skip creating a Controller Service if one was
not needed, and if it was needed, the service would would be created at
the end of BeforeServe. With GoCSI, this was too late because a check
was done to see that the ControllerService was not nil, and the previous
code told GoCSI that it was nil.

This patch changes GetController() to instantiate a ControllerService
struct early and return it, even though it is not initialize. To
facilitate this, New() not returns a nil struct, and Init() has been
added to actually configure the controller with the vSphere config file.

This patch also removes the calls for `log.Fatal*` - we need to avoid
the use of Fatal as it prevents GoCSI from performing automatic sock
file cleanup.