Optimize cross-build/cross-dist targets
This patch optimizes the way the cross-build and cross-dist targets are
handled. Recursive make invocations do not define the targets and the
dependency graph of a target's sources is not regenerated in recursive
calls either. Normally this would be an issue due to the way different
GOOS and GOARCH values can affect which sources are used, but "find" is
used to get a list of the Go files, not "go list", and so the initial
list should be close to comprehensive.