Touch the CCM and CSI binaries on build
This patch makes sure the CCM and CSI binaries are touched when built.
This is because Go may not actually rebuild the binaries if nothing has
changed (Go 1.10+ supports caching for the "go build" command). This can
present an issue if the reason the target was rebuilt doesn't result in
a binary with an M_TIME newer than the dependency that caused the build.
The "touch" command ensures the binaries' M_TIMEs are always newer than
whatever dependency caused the build to occur.