ci: CI image for Prow
This patch introduces the image "gcr.io/cloud-provider-vsphere/ci", used
as the image for the CCM and CSI provider's Prow jobs. This image
contains the Go module and build caches from a recent build of the
project, therefore the image can build the binaries in a matter of
moments.

Plesae refer to the documentation at "hack/images/ci/README.md" for
instructions on how to use this image locally to mimic all of the Prow
jobs' workflows.

This patch also deletes the image "gcr.io/cloud-provider-vsphere/golang"
in favor of using the new "ci" image as the basis for quick builds.