ci: Rm Makefile from CI build cache; LDFLAGS
This patch removes the need to copy the Makefile when creating the build
cache for the CI image in order to not invalidate the cache due to a
Makefile change. To support this the LDFLAGS have been externalized to
"hack/make/ldflags.txt" so the Makefile and CI Dockerfile both use the
same flags.