Introduce new release tooling and structure
This patch adds a hack/release.sh script that can be used both locally
and in CI for generating and pushing images and binaries to GCR and GCS.
The Makefile is modified to make use of this script, and the integration
test is modified to use the new images.

The location of the images and binaries moves with this change, mainly
splitting the latest builds from master into one channel versus a
release channel that only contains tagged releases. Doc updates and
manifest changes need to come in after this as well.

The CPI container image is also now based on distroless.