Always ensure KIND is present for integration test
This patch adds a step in the integration test Makefile to explicitly
install KIND if it is not present. The default KIND version is now
v0.5.1, which requires changes to the kind config to use the v1alpha3
API.

Part of the movitivation behind this is to be able to run the
integration test in the kubekins-e2e image rather than a custom "CI"
image created from this project.