Removed references to 'official' from README
There were some internal queries about the use of the word 'official' in
this repo with reference to CSI from a documentation perspective. The
'official' source is docs.vmware.com, however this repo is more up to
date. We've been asked to remove the reference to official, net effect
other than the verbiage change is nothing.