instances: skip node deletion by VM name
Skip node deletion by VM name since a VM will not exist according to vCenter between the time the node reports itself to the apiserver (as NotReady) and when vmtools is correctly reporting the hostname. By only deleting a VM by UUID, we ensure only existing registered nodes can be deleted when it no longer exists. The caveat here is that cloud provider will no longer delete nodes that failed to register and no longer exist, but this is much better than accidentally deleting a VM where vmtools was slow to report the hostname

Signed-off-by: Andrew Sy Kim <kim.andrewsy@gmail.com>