Return error if multiple VMs with same hostnames found
In case if multiple VMs with the same hostname exists
within the same scope, picking wrong VM is quite possible,
which might lead to hard-diagnosable issues with
Nodes initialization, or even to incorrect Node initialization.

This commit replaces `FindByDnsName` call to `FindAllByDnsName`,
`GetVMByDNSName` function now returns an error in case if
multiple VMs with the same HostName was found.