CredentialsManager accepts alternate Secret format
This change was made because the legacy Secret format is not compatible
with IPv6 server addresses. The legacy format uses the server address in
the Secret key and because IPv6 addresses contain colons, errors
occur.

This change continues to support the legacy Secret format.

This change adds the ability to add Secrets in the following format:

```yaml
---
apiVersion: v1
kind: Secret
metadata:
  name: vsphere-cloud-secret
  namespace: kube-system
data:
  10.96.96.96.username: "dXNlcg=="
  10.96.96.96.password: "cGFzcw=="

  username_0: "dXNlcg=="
  password_0: "cGFzcw=="
  server_0: "ZmQwMDo6MQo="

  username_foo: "dXNlcg=="
  password_foo: "cGFzcw=="
  server_foo: "ZmQwMDo6Mgo="
```

The username_, password_, server_ keys should be supplied with matching suffixes.

Signed-off-by: Tyler Schultz <tschultz@vmware.com>
Co-authored-by: Aidan Obley <aobley@vmware.com>